requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-celery-beat==2.5.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.3
flower==2.0.1
libvirt-python==9.0.0
django-filter==23.5
drf-spectacular==0.27.1
python-decouple==3.8
django-extensions==3.2.3
